+++ src/devices/bus/ata/ataintf.h
@@ -57,7 +57,7 @@ public:
 
 	ata_slot_device &slot(int index);
 
-	uint32_t read_dma();
+	uint16_t read_dma();
 	void write_dma(uint16_t data);
 	DECLARE_WRITE_LINE_MEMBER(write_dmack);
